+++ src/remotior_sensus/tools/band_clip.py
@@ -36,7 +36,7 @@ from remotior_sensus.core import configurations as cfg
 from remotior_sensus.core.bandset_catalog import BandSet
 from remotior_sensus.core.bandset_catalog import BandSetCatalog
 from remotior_sensus.core.output_manager import OutputManager
-from remotior_sensus.util.files_directories import move_file, parent_directory
+from remotior_sensus.util.files_directories import move_file
 from remotior_sensus.util.shared_tools import prepare_process_files
 from remotior_sensus.util.raster_vector import get_vector_values
 from remotior_sensus.core.processor_functions import clip_raster